requirements.txt
absl-py==1.3.0
asgiref==3.5.2
astunparse==1.6.3
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
colorama==0.4.6
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
dj-database-url==1.0.0
Django==3.2.15
django-bootstrap4==22.2
django-filter==22.1
django-heroku==0.3.1
django-import-export==2.8.0
django-jazzmin==2.5.0
django-phonenumber-field==6.4.0
django-tables2==2.5.3
djangorestframework==3.13.1
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.8.0
fonttools==4.38.0
google-pasta==0.2.0
gunicorn==20.1.0
h5py==3.7.0
importlib-metadata==5.0.0
ipython==7.34.0
jedi==0.18.2
joblib==1.2.0
jupyter-client==7.4.9
jupyter-core==4.12.0
kiwisolver==1.4.4
MarkupPy==1.14
matplotlib==3.5.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.21.6
odfpy==1.4.1
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2022.10.25
platformdirs==2.5.2
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.3
pycodestyle==2.10.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
seaborn==0.12.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
tensorflow-io-gcs-filesystem==0.31.0
tomli==2.0.1
tornado==6.2
traitlets==5.8.1
typing-extensions==4.3.0
tzdata==2022.2
virtualenv==20.16.4
virtualenv-clone==0.5.7
wcwidth==0.2.6
whitenoise==6.2.0
widgetsnbextension==4.0.5
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
